emulators/ares: update to 142, change to CMake
Release notes: https://ares-emu.net/news/ares-v142-released

Upstream overhauled the build system and changed to CMake. The old GNU make
based build system is still available and working, but it is declared deprecated
and will be removed in the next release. Therefore the port changes to the new
CMake based build system now.

The port reflects the change to CMake by removing, splitting, combining and
adding some port options, and it allows to remove most of the custom targets and
patches. Overall it simplifies and streamlines the port in many parts.

Previously, linking against libopenal.so with lld failed on i386. LLD_UNSAFE=yes
was used as a workaround. Now the new mandatory tools/sourcery target chokes on
it with enabled LTO:

  /usr/local/bin/i386-unknown-freebsd14.1-ld.bfd: /usr/bin/../lib/LLVMgold.so:
  error loading plugin: Cannot open "/usr/bin/../lib/LLVMgold.so"

There are plenty of other audio backend options available, especially the OSS
backend works perfectly. In lieu of a proper fix and LTO being more important,
in particular for demanding cores like the N64, the port excludes the OPENAL
option on i386 now.

- change to CMake
- add mandatory devel/librashader to BUILD_DEPENDS
- add mandatory gl to USES
- remove obsolete localbase:ldflags from USES
- exclude unbundled dependencies from extraction via EXTRACT_AFTER_ARGS
- add option TOOLS to include supplemental tools, disabled by default
- split option SHADER into LIBRASHADER and LIBRETRO
- combine options ACCURACY and PERFORMANCE into boolean ACCURACY
- combine options SDLAUDIO and SDLINPUT into SDL
- remove obsolete options DEBUG, DOCS, GLX, OPENMP, XLIB and XVIDEO
- rename option UHID to USBHID
- exclude option OPENAL on i386
- update *_DESC
- simplify the HOTKEYS extra patch
- remove custom do-install target
- remove obsolete post-patch-SHADER-on and post-install-DOCS-on targets
- simplify location of shaders by setting a symbolic link to contents of
  games/libretro-shaders-slang in post-install-LIBRETRO-on target instead of
  using a custom patch and post-patch target
- remove obsolete patches
- improve pkg-message