games/libretro-shaders-slang: update to g20241122
Commitlog:
https://github.com/libretro/slang-shaders/compare/4f3d214b314d84c6a3e25b27c77bd1d0ee3a1697..8731861b22946faf71bb94ce25147484db61eb6f

Upstream's Makefile has become usable without patches and deletes the unwanted
files now. This allows us to simplify the port by removing EXTRACT_AFTER_ARGS
and the custom do-install target.

There are documentation files installed into the DATADIR tree, which we do not
delete or move away. This would be an unneeded and fruitless burden maintaining
the port, so it was never done. In order to keep the port simple, the top-level's
README.md is now also treated this way again and the DOCS port option is removed.
Otherwise, custom pre-install-DOCS-on/-off targets would be required to handle
this file appropriatly, which complicates the port without real benefit.